requirements.txt
binance-connector
prometheus-client
coverage
uwsgi
responses